Use ensure_safe_goal_position

+++ src/lerobot/robots/reachy2/configuration_reachy2.py
@@ -25,9 +25,8 @@ from ..config import RobotConfig
 @dataclass
 class Reachy2RobotConfig(RobotConfig):
     # `max_relative_target` limits the magnitude of the relative positional target vector for safety purposes.
-    # Set this to a positive scalar to have the same value for all motors, or a list that is the same length as
-    # the number of motors in your follower arms.
-    max_relative_target: int | None = None
+    # Set this to a positive scalar to have the same value for all motors.
+    max_relative_target: float | None = None
 
     # IP address of the Reachy 2 robot
     ip_address: str | None = "localhost"
